+++ Src/IronPython/Runtime/Bytes.cs
@@ -856,7 +856,9 @@ namespace IronPython.Runtime {
         }
 
         public virtual string __str__(CodeContext context) {
-            PythonOps.Warn(context, PythonExceptions.BytesWarning, "str() on a bytes instance");
+            if (context.LanguageContext.PythonOptions.BytesWarning) {
+                PythonOps.Warn(context, PythonExceptions.BytesWarning, "str() on a bytes instance");
+            }
             return _bytes.BytesRepr();
         }
